refactored code, added "good bot" reaction, .env files
