+++ .env.example
@@ -1,0 +1,1 @@
+BOT_TOKEN=test
