+++ .gitignore
@@ -1,3 +1,4 @@
 /.idea
 vendor
 .phpunit.result.cache
+.env
