+++ composer.json
@@ -12,7 +12,8 @@
   "require": {
     "react/event-loop": "^1.3",
     "react/http": "^1.7",
-    "team-reflex/discord-php": "^7.1"
+    "team-reflex/discord-php": "^7.1",
+    "vlucas/phpdotenv": "^5.5"
   },
   "require-dev": {
     "phpunit/phpunit": "^9.5"
